diary/20180918/requirements.txt
Pillow==5.2.0
requests==2.19.1
